Moving ES ID generator into generators package + singleton construction
