+++ libbeat/processors/uuid/generator/mac.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package elasticsearch
+package generator
 
 import (
 	"crypto/rand"
